+++ inst/www/js/app_esttest.js
@@ -10,19 +10,19 @@ $(document).ready(function(){
 		sessionStorage.removeItem('generatedData');
 		sessionStorage.removeItem('dataCenter');
 	});
-	
+
 	$("#picture").load(function() {
 		var canvas_fg = $("#canvas_fg");
 		var canvas_bg = $("#canvas_bg");
-        context_fg = canvas_fg[0].getContext('2d');
+    context_fg = canvas_fg[0].getContext('2d');
 		context_bg = canvas_bg[0].getContext('2d');
-		
+
 		context_fg.canvas.height = $(this).height();
-        context_fg.canvas.width = $(this).width();
+    context_fg.canvas.width = $(this).width();
 		context_bg.canvas.height = $(this).height();
-        context_bg.canvas.width = $(this).width();
+    context_bg.canvas.width = $(this).width();
 	});
-	
+
 	$("#err_deviation").hide();
 	$("#var_dev").change(function(){
 		if ($(this).val() == "value")
@@ -30,20 +30,20 @@ $(document).ready(function(){
 		else
 			$("#err_deviation").hide();
 	});
-	
+
 	$(".inputparam").change(function(){
 		if (!$("#endpoint").hasClass("invisible")) {
 			$("#submitajax_est, #submitajax_test").prop("disabled", true);
 		}
 	});
-	
+
 	$("#canvas_fg").mousedown(function(event) {
 		if ($("#picture").attr("src") == "#") {
 			return;
 		}
-		
+
 		var ratio = $("#picture").width() / $("#picture")[0].naturalWidth;
-		
+
 		if ($("#coordx1").is(':empty')) {
 			$("#coordx1").text(((event.pageX - $(this).offset().left) / ratio).toFixed()) ;
 			$("#coordy1").text(((event.pageY - $(this).offset().top) / ratio).toFixed()) ;
@@ -52,19 +52,19 @@ $(document).ready(function(){
 			$("#coordx2").text(((event.pageX - $(this).offset().left) / ratio).toFixed());
 			$("#coordy2").text(((event.pageY - $(this).offset().top) / ratio).toFixed()) ;
 			$("#endpoint").removeClass("invisible");
-			
+
 			$("#submitajax_prepare").removeAttr("disabled");
-			
+
 			// draw a line
 			var x1 = parseInt($("#coordx1").text())*ratio;
 			var y1 = parseInt($("#coordy1").text())*ratio;
 			var x2 = parseInt($("#coordx2").text())*ratio;
 			var y2 = parseInt($("#coordy2").text())*ratio;
-			
+
 			createLine(x1,y1, x2,y2, "LightGreen", "canvas_bg");
 		}
 	});
-	
+
 	$("#resetbutton").click(function(){
 		clearCanvas("canvas_bg");
 		clearCanvas("canvas_fg");
@@ -74,35 +74,35 @@ $(document).ready(function(){
 		sessionStorage.removeItem('generatedData');
 		sessionStorage.removeItem('dataCenter');
 	});
-	
+
 	$("#resetoutputbutton").click(function(){
 		$("#status").html("<span class='tempoutput'>Status: waiting....</span>");
 		$("#data_plot").empty();
 	});
-		
+
 	$("#submitajax_prepare, #submitajax_est, #submitajax_test").click(function(){
 		$(".tempoutput").remove();
 		if (!$("#status").is(':empty'))
 			$("#status").prepend("<hr>");
 		$("#status").prepend("<span class='tempoutput'><img src='pictures/spinner.gif'> Calculation in progress...<br></span>");
 		$("#resetoutputbutton").addClass("disabled");
-				
+
 		if (this.id == "submitajax_prepare") {
 			testtype = "prepare";
 			sessionStorage.removeItem('generatedData');
 			sessionStorage.removeItem('dataCenter');
 		} else if (this.id == "submitajax_est") {
-			testtype = "est";			
-		} else if (this.id == "submitajax_test") {			
+			testtype = "est";
+		} else if (this.id == "submitajax_test") {
 			testtype = "test";
 		}
-		
+
 		var generateddata = 0, datacenter = 0;
 		if (testtype == "est" || testtype == "test") {
 			generateddata = sessionStorage.generatedData;
 			datacenter = sessionStorage.dataCenter;
 		}
-		
+
 		var req = ocpu.call("load.image", {
 			generatedData : generateddata,
 			dataCenter : datacenter,
@@ -128,46 +128,47 @@ $(document).ready(function(){
 		}, function(session) {
 			session.getObject(function(outtxt){
 				$("#status").prepend(outtxt[0]);
-				
-				if (testtype = "prepare") {
+
+				if (testtype == "prepare") {
 					$("#submitajax_est, #submitajax_test").removeAttr("disabled");
 				}
-				
+
 				var ratio = $("#picture").width() / $("#picture")[0].naturalWidth;
-				
+
 				x1 = parseInt(outtxt[2])*ratio;
 				y1 = parseInt(outtxt[1])*ratio;
 				x2 = parseInt(outtxt[4])*ratio;
 				y2 = parseInt(outtxt[3])*ratio;
-				
+
 				clearCanvas("canvas_fg");
 				createLine(x1, y1, x2, y2, "Red", "canvas_fg");
-				
+
 				// Generated data. Must be saved in browser local storage.
 				if (!sessionStorage.generatedData) {
 					sessionStorage.generatedData = outtxt[5];
 					sessionStorage.dataCenter = outtxt[6];
 				}
 			});
-			
+
 			session = session.getKey();
 			$(".tempoutput").remove();
-			$("#status").prepend("<br><a target='_blank' href='/ocpu/tmp/"+ session +"/console'>Click for raw R output</a><br>")
-			
+			$("#status").prepend("<br><a target='_blank' href='/ocpu/tmp/"+ session +"/console'>Click for raw R output</a><br>");
+
 			if (testtype == "prepare") {
-				$("#data_plot").html("<img src='/ocpu/tmp/"+ session +"/graphics/last/png'>");
+				// $("#data_plot").html("<img src='/ocpu/tmp/"+ session +"/graphics/last/png'>");
+				$("#status").prepend("<img src='/ocpu/tmp/"+ session +"/graphics/last/png'>");
 			}
-			
+
 			$("#resetbutton").removeClass("disabled");
 			$("#resetoutputbutton").removeClass("disabled");
 		});
-		
+
 		req.fail(function(){
 			clearCanvas("canvas_fg");
 			$(".tempoutput").remove();
 			$("#status").prepend("R returned an error: " + req.responseText);
 			$("#resetbutton").removeClass("disabled");
 			$("#resetoutputbutton").removeClass("disabled");
-		});	
+		});
 	});
 });